build: fix issue with file perms on Linux
When building in Docker on Linux, must take
care of file ownership for files produces within
the running container. By default they will be
root, which messes up a lot of things.

For Linux only we run as the launching user
with flags `-u $(id -u):$(id -g)`.

The issue does not exist on OS X.
Not sure if the issue exists on Windows.